Refactor into separate files and reset hero isEnabled status after dismissal
